+++ index.py
@@ -1,14 +1,15 @@
 import os
 
-from flask import Flask, url_for, render_template, request, send_from_directory
+from flask import Flask, jsonify, url_for, render_template, request, send_from_directory
 from flask_basicauth import BasicAuth
 from flask_sqlalchemy import SQLAlchemy
+from helpers import get_env
+
 
 app = Flask(__name__)
 
-app.config['BASIC_AUTH_USERNAME'] = 'wtpa'
-app.config['BASIC_AUTH_PASSWORD'] = 'l0g5'
-
+app.config['BASIC_AUTH_USERNAME'] = get_env('BASIC_AUTH_USERNAME')
+app.config['BASIC_AUTH_PASSWORD'] = get_env('BASIC_AUTH_PASSWORD')
 app.config['DEBUG'] = True
 
 basic_auth = BasicAuth(app)
@@ -36,29 +37,16 @@ def argst():
 @app.route('/logs/')
 @basic_auth.required
 def logs():
-    import os
-    oot_dir = os.path.dirname(os.getcwd())
-
-    return send_from_directory(os.path.join(root_dir, 'static', 'irc_logs'), '2018-06-27.txt')
+    resp = { 'status': 123 }
+    return jsonify(resp)
 
 
-@app.route('/test/')
-def test():
-    url = get_env('POSTGRES_URL')
-    return render_template('test.html', url=url)
 
+#@app.route('/test/')
+#def test():
+#    url = get_env('POSTGRES_URL')
+#    return render_template('test.html', url=url)
 
-# ============
-# HELPERS
-# ============
-
-# loaded from .env file in index.wsgi
-def get_env(env_name):
-    try:
-        return os.environ[env_name]
-    except KeyError:
-        msg = "Variable '{}' not found".format(env_name)
-        raise Exception(msg)
