--- test.py
@@ -1,5 +1,0 @@
-@app.route('/test/')
-def argst():
-    return render_template('test.html')
-
-
